fix(printf): Add low-address pointer protection to fmtstrcpy
Root cause of WASM3 hang was corrupted pointers passed to printf %s format.
fmtstrcpy() already had NULL protection but not for low-address pointers.

Changes:
- Added check for (uintptr)s < 0x1000 in fmtstrcpy() to treat as NULL
- Fixed %lu to %d in clr_runtime.c debug prints (kernel printf limitation)
- Removed calloc debug prints that used unsupported %lu format

Progress:
- m3_NewEnvironment() now works
- m3_NewRuntime() now works
- m3_ParseModule() still under investigation (hangs inside WASM3)

+++ kernel/clr/clr_runtime.c
@@ -199,7 +199,7 @@ int clr_execute_assembly(void *dll_data, ulong dll_size) {
   }
 
   print("CLR: Calling m3_ParseModule...\n");
-  print("CLR: DEBUG: wasm_bytes=%p wasm_size=%lu\n", wasm_bytes, wasm_size);
+  print("CLR: DEBUG: wasm_bytes=%p wasm_size=%d\n", wasm_bytes, (int)wasm_size);
   IM3Module module = NULL;
   M3Result result = m3_ParseModule(env, &module, wasm_bytes, (u32int)wasm_size);
   print("CLR: m3_ParseModule() returned result=%p\n", result);
